🐛 Separate refresh logic to avoid weird bundled code in vitest
